refactor(sources): modularize Discord kernel
